automation.elastic.ilm_excel/requirement.txt
openpyxl==3.1.2
pandas==2.2.2
pyyaml==6.0.1
requests==2.31.0